backend/requirements.txt
# OpenAI SDK with Responses API support (GPT-5)
openai>=1.55.0

# FastAPI and WebSocket
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
websockets>=13.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9

# Pydantic for structured outputs
pydantic>=2.9.0

# Utilities
python-dotenv>=1.0.0
